day5_production/requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
flask>=2.3.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
cachetools>=5.3.0
aiohttp>=3.9.0

